pyproject.toml
[tool.poetry]
name = "brightness-comparator"
version = "0.1.0"
description = ""
authors = ["megahomyak <g.megahomyak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
pynput = "^1.7.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
